requirements.txt
streamlit
pandas
fpdf
rapidfuzz
google-generativeai
groq
colorama